+++ flink-formats/flink-csv/src/main/java/org/apache/flink/formats/csv/CsvFormatFactory.java
@@ -48,9 +48,9 @@ import static org.apache.flink.formats.csv.CsvOptions.ARRAY_ELEMENT_DELIMITER;
 import static org.apache.flink.formats.csv.CsvOptions.DISABLE_QUOTE_CHARACTER;
 import static org.apache.flink.formats.csv.CsvOptions.ESCAPE_CHARACTER;
 import static org.apache.flink.formats.csv.CsvOptions.FIELD_DELIMITER;
-import static org.apache.flink.formats.csv.CsvOptions.IGNORE_PARSE_ERRORS;
 import static org.apache.flink.formats.csv.CsvOptions.NULL_LITERAL;
 import static org.apache.flink.formats.csv.CsvOptions.QUOTE_CHARACTER;
+import static org.apache.flink.table.factories.FactoryUtil.IGNORE_PARSE_ERRORS;
 
 /**
  * Format factory for providing configured instances of CSV to RowData {@link SerializationSchema}
