+++ flink-formats/flink-json/src/main/java/org/apache/flink/formats/json/canal/CanalJsonOptions.java
@@ -22,13 +22,12 @@ import org.apache.flink.configuration.ConfigOption;
 import org.apache.flink.configuration.ConfigOptions;
 import org.apache.flink.configuration.ReadableConfig;
 import org.apache.flink.formats.json.JsonOptions;
+import org.apache.flink.table.factories.FactoryUtil;
 
 /**
  * Option utils for canal-json format.
  */
 public class CanalJsonOptions {
-
-	public static final ConfigOption<Boolean> IGNORE_PARSE_ERRORS = JsonOptions.IGNORE_PARSE_ERRORS;
 
 	public static final ConfigOption<String> TIMESTAMP_FORMAT = JsonOptions.TIMESTAMP_FORMAT;
